Move i18n related files to i18n folder
