+++ pages/home/index.spec.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { shallow, mount } from 'enzyme';
 import HomePage from '../home/index';
 import { I18nextProvider } from 'react-i18next';
-import i18nForTests from '../../app/i18nForTests';
+import i18nForTests from '../../app/i18n/i18nForTests';
 
 import { Provider } from 'react-redux';
 
@@ -15,8 +15,6 @@ describe('HomePage', () => {
 	});
 
 	it('should render without throwing an error', () => {
-		i18nForTests.language = 'en';
-
 		const wrap = mount(
 			<Provider store={initStore()}>
 				<I18nextProvider i18n={i18nForTests}>
